Move uninstall to singleton manager
