+++ micropip/list.py
@@ -8,24 +8,6 @@ from .package import PackageDict, PackageMetadata
 def list_installed_packages(
     lockfile_packages: dict[str, dict[str, Any]]
 ) -> PackageDict:
-    """Get the dictionary of installed packages.
-
-    Returns
-    -------
-    ``PackageDict``
-        A dictionary of installed packages.
-
-        >>> import micropip
-        >>> await micropip.install('regex') # doctest: +SKIP
-        >>> package_list = micropip.list()
-        >>> print(package_list) # doctest: +SKIP
-        Name              | Version  | Source
-        ----------------- | -------- | -------
-        regex             | 2021.7.6 | pyodide
-        >>> "regex" in package_list # doctest: +SKIP
-        True
-    """
-
     # Add packages that are loaded through pyodide.loadPackage
     packages = PackageDict()
     for dist in importlib.metadata.distributions():
